+++ ES Game Library/MakeClasses/Fujimura/Judgement/LongNoteJudgement.h
@@ -10,7 +10,7 @@ public:
 	LongNoteJudgement();
 	~LongNoteJudgement();
 
-	virtual JUDGE Judge(Note* judgeNote, DWORD nowTime, RawInputMouse& mouse);
+	virtual JUDGE Judge(Note* judgeNote, LONG nowTime, RawInputMouse& mouse);
 
 private:
